requirements.txt
# Dosya: requirements.txt
# Açıklama: Render deployment için Python bağımlılıkları - SAM + YOLO

Flask==2.3.3
Flask-CORS==4.0.0
opencv-python-headless==4.8.1.78
numpy==1.26.4
gunicorn==21.2.0
setuptools==69.5.1
wheel==0.43.0
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196
segment-anything==1.0
Pillow==10.0.1
